Add aspire models for resources and properties.
The Aspire project uses a good practice of defining well-known properties and
resource types in a single place. This commit adds the `KnownProperties` and
`KnownResourceTypes` classes to the `Aspire.ResourceService.Standalone.Server` project.

It also adds a `GlobalSuppressions` file to suppress warnings about the different
namespace used in the imported files.