Add resource notification service
This service allows publishing and subscribing
to changes in he state of a resource.
Essentially, this would be the service that resource providers
call to inform others of a change in the state of a resource.
The `DashboardData` will be responsible for subscribing to
changes published to the `ResourceNotificationService` and
publishing them to Resource Reporterso that others get notified
of the change.